pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[project]
authors = [
  {name = "gaetan"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "ipykernel>=7.1.0",
  "kaggle>=1.8.2",
  "loguru",
  "matplotlib>=3.10.8",
  "mkdocs",
  "joblib>=1.3.0",
  "pillow>=10.0.0",
  "pip",
  "plotly>=6.5.0",
  "polars>=1.36.1",
  "pytest",
  "python-dotenv",
  "ruff",
  "torch>=2.1.0",
  "tqdm",
  "transformers>=4.38.0",
  "typer",
  "xgboost>=2.0.0",
  "nbformat>=4.2.0",
  "scikit-learn>=1.8.0",
]
description = "A short description of the project."
license = {file = "LICENSE"}
name = "image2biomass"
readme = "README.md"
requires-python = "~=3.13.0"
version = "0.0.1"

[tool.ruff]
include = ["pyproject.toml", "image2biomass/**/*.py"]
line-length = 99
src = ["image2biomass"]

[tool.ruff.lint]
extend-select = ["I"] # Add import sorting

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["image2biomass"]
